CREATE interfaces: Single Reposonsibility part 2.
Interfaces tightly coupled with Model class type as they require TModle
type to be Model<,>

IFirstModel<TModel, TID: Represents an object which offers
the first model with in its internal collection.

IReadable<TModelDTO, TModel, TID>: Represents an object that allows
reading a single model or multiple models.
(Only if not MODEL_NONREADABLE)

IDeletable<TModelDTO, TModel, TID>: Represents an object that allows
deleting a single model with a specified ID.
(Only if MODEL_DELETABLE)

IAppendable<TModelDTO, TModel, TID>: Represents an object that has
a list of models to which a new model can be appended.
(Only if MODEL_APPENDABLE)

IUpdatable<TModelDTO, TModel, TID>: represents an object that has a list
of models and allows a model with a specified ID to be updated with data
from the given model parameter.
(Only if MODEL_UPDATABLE)