CREATE interfaces: Single Resonsibility part:3 For IActionResult support: IReadable<TModel, TID>: Represents an object that allows reading a single model or multiple models. (Only if not MODEL_NONREADABLE)
IDeletable<TModel, TID>: Represents an object that allows
deleting a single model with a specified ID.
(Only if MODEL_DELETABLE)

IAppendable<TModel, TID>: Represents an object that has
a list of models to which a new model can be appended.
(Only if MODEL_APPENDABLE)

IUpdatable<TModel, TID>: represents an object that has a list
of models and allows a model with a specified ID to be updated with data
from the given model parameter.
(Only if MODEL_UPDATABLE)